Fix compiler warning

+++ Objects/longobject.c
@@ -1335,7 +1335,7 @@ PyLong_FromLongLong(long long ival)
     /* Count digits (at least two - smaller cases were handled above). */
     abs_ival = ival < 0 ? 0U-(unsigned long long)ival : (unsigned long long)ival;
     /* Do shift in two steps to avoid possible undefined behavior. */
-    t = abs_ival >> PyLong_SHIFT >> PyLong_SHIFT;
+    t = (Py_ssize_t)(abs_ival >> PyLong_SHIFT >> PyLong_SHIFT);
     ndigits = 2;
     while (t) {
         ++ndigits;
